requirements.txt
streamlit>=1.37
torch>=2.4
torchvision>=0.19
timm>=1.0.7
pillow
numpy
# ถ้าเช็คพอยต์เคยเซฟผ่าน lightning/fabric:
lightning>=2.4
huggingface_hub>=0.23
